moved ChartGroup router and rest handler to respective folder
