+++ api/appStore/chartGroup/ChartGroupRouter.go
@@ -15,26 +15,25 @@
  *
  */
 
-package router
+package chartGroup
 
 import (
-	"github.com/devtron-labs/devtron/api/restHandler"
 	"github.com/gorilla/mux"
 )
 
 type ChartGroupRouterImpl struct {
-	ChartGroupRestHandler restHandler.ChartGroupRestHandler
+	ChartGroupRestHandler ChartGroupRestHandler
 }
 type ChartGroupRouter interface {
-	initChartGroupRouter(helmRouter *mux.Router)
+	InitChartGroupRouter(helmRouter *mux.Router)
 }
 
-func NewChartGroupRouterImpl(ChartGroupRestHandler restHandler.ChartGroupRestHandler) *ChartGroupRouterImpl {
+func NewChartGroupRouterImpl(ChartGroupRestHandler ChartGroupRestHandler) *ChartGroupRouterImpl {
 	return &ChartGroupRouterImpl{ChartGroupRestHandler: ChartGroupRestHandler}
 
 }
 
-func (impl *ChartGroupRouterImpl) initChartGroupRouter(chartGroupRouter *mux.Router) {
+func (impl *ChartGroupRouterImpl) InitChartGroupRouter(chartGroupRouter *mux.Router) {
 	chartGroupRouter.Path("/").
 		HandlerFunc(impl.ChartGroupRestHandler.CreateChartGroup).Methods("POST")
 	chartGroupRouter.Path("/").
